requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# ICARUS-X  —  Space Weather Forecasting System
# Python 3.11 pinned dependencies
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ── Core ML ──────────────────────────────────────────────
torch>=2.2.2
torchvision>=0.17.2
numpy>=1.26.4
scipy>=1.13.0
pandas>=2.2.2
scikit-learn>=1.4.2

# ── XGBoost (M3 Sentinel) ───────────────────────────────
xgboost==2.0.3
imbalanced-learn==0.12.2        # SMOTE oversampling
shap==0.45.1                    # SHAP explainability

# ── YOLOv10 (M1 Visionary — plugged in tomorrow) ────────
ultralytics==8.2.18

# ── API / Backend (M5 Architect) ─────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Database ─────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0                 # async PostgreSQL driver
psycopg2-binary==2.9.9          # sync PostgreSQL driver
alembic==1.13.1

# ── HTTP / Data Fetching ─────────────────────────────────
httpx==0.27.0                   # async HTTP for NOAA poller
aiofiles==23.2.1

# ── Monitoring / Logging ─────────────────────────────────
loguru==0.7.2
tqdm==4.66.4

# ── Utilities ────────────────────────────────────────────
python-dateutil==2.9.0.post0
joblib==1.4.2
matplotlib==3.8.5               # for training plots
